requirements.txt
asgiref==3.7.2
Django==4.1.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
Faker==19.3.1
mongomock==4.1.2
packaging==23.1
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
sentinels==1.0.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3